# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API COVID-19 et Monkeypox avec gestion des Continents et Pays"
  version: "1.0.0"
  title: "API COVID-19 & Monkeypox"
host: "localhost:3002" 
basePath: "/"
schemes:
  - "http"
...
paths:
  /continents:
    get:
      tags: [Continents]
      summary: "Récupère la liste de tous les continents"
      description: "Retourne toutes les entrées de la table 'continent'."
      responses:
        200:
          description: "Liste des continents"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Continent"
    post:
      tags: [Continents]
      summary: "Crée un nouveau continent"
      description: "Ajoute une nouvelle entrée dans la table 'continent'."
      parameters:
        - in: "body"
          name: "body"
          description: "Objet continent à ajouter"
          required: true
          schema:
            $ref: "#/definitions/Continent"
      responses:
        201:
          description: "Continent créé avec succès"
          schema:
            $ref: "#/definitions/Continent"
        400:
          description: "Requête invalide"
  /continents/{idContinent}:
    get:
      tags: [Continents]
      summary: "Récupère un continent par son ID"
      description: "Retourne les détails d'un continent spécifique."
      parameters:
        - name: "idContinent"
          in: "path"
          description: "ID du continent à récupérer"
          required: true
          type: "integer"
      responses:
        200:
          description: "Détails du continent"
          schema:
            $ref: "#/definitions/Continent"
        404:
          description: "Continent non trouvé"
    put:
      tags: [Continents]
      summary: "Met à jour un continent existant"
      description: "Met à jour les détails d'un continent par son ID."
      parameters:
        - name: "idContinent"
          in: "path"
          description: "ID du continent à mettre à jour"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Objet continent avec les données à mettre à jour"
          required: true
          schema:
            $ref: "#/definitions/Continent"
      responses:
        200:
          description: "Continent mis à jour avec succès"
        400:
          description: "Requête invalide"
        404:
          description: "Continent non trouvé"
    delete:
      tags: [Continents]
      summary: "Supprime un continent"
      description: "Supprime un continent par son ID."
      parameters:
        - name: "idContinent"
          in: "path"
          description: "ID du continent à supprimer"
          required: true
          type: "integer"
      responses:
        200:
          description: "Continent supprimé avec succès"
        404:
          description: "Continent non trouvé"
  /pays:
    get:
      tags: [Pays]
      summary: "Récupère la liste de tous les pays"
      description: "Retourne toutes les entrées de la table 'pays', incluant le nom du continent."
      responses:
        200:
          description: "Liste des pays"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Pays"
    post:
      tags: [Pays]
      summary: "Crée un nouveau pays"
      description: "Ajoute une nouvelle entrée dans la table 'pays'."
      parameters:
        - in: "body"
          name: "body"
          description: "Objet pays à ajouter"
          required: true
          schema:
            $ref: "#/definitions/PaysInput" # Utilisation d'un schéma d'input pour éviter l'ID généré
      responses:
        201:
          description: "Pays créé avec succès"
          schema:
            $ref: "#/definitions/Pays"
        400:
          description: "Requête invalide"
  /pays/{id_pays}:
    get:
      tags: [Pays]
      summary: "Récupère un pays par son ID"
      description: "Retourne les détails d'un pays spécifique, incluant le nom du continent."
      parameters:
        - name: "id_pays"
          in: "path"
          description: "ID du pays à récupérer"
          required: true
          type: "integer"
      responses:
        200:
          description: "Détails du pays"
          schema:
            $ref: "#/definitions/Pays"
        404:
          description: "Pays non trouvé"
    put:
      tags: [Pays]
      summary: "Met à jour un pays existant"
      description: "Met à jour les détails d'un pays par son ID."
      parameters:
        - name: "id_pays"
          in: "path"
          description: "ID du pays à mettre à jour"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Objet pays avec les données à mettre à jour"
          required: true
          schema:
            $ref: "#/definitions/PaysInput"
      responses:
        200:
          description: "Pays mis à jour avec succès"
        400:
          description: "Requête invalide"
        404:
          description: "Pays non trouvé"
    delete:
      tags: [Pays]
      summary: "Supprime un pays"
      description: "Supprime un pays par son ID."
      parameters:
        - name: "id_pays"
          in: "path"
          description: "ID du pays à supprimer"
          required: true
          type: "integer"
      responses:
        200:
          description: "Pays supprimé avec succès"
        404:
          description: "Pays non trouvé"
  /covid_country:
    get:
      tags: [CovidCountry]
      summary: "Récupère toutes les données COVID par pays"
      description: "Retourne toutes les entrées de la table 'covid_country', incluant les noms du pays et du continent."
      responses:
        200:
          description: "Données COVID par pays"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CovidCountry"
    post:
      tags: [CovidCountry]
      summary: "Ajoute de nouvelles données COVID pour un pays"
      description: "Ajoute une nouvelle entrée dans la table 'covid_country'."
      parameters:
        - in: "body"
          name: "body"
          description: "Données COVID à ajouter"
          required: true
          schema:
            $ref: "#/definitions/CovidCountryInput"
      responses:
        201:
          description: "Donnée ajoutée avec succès"
          schema:
            $ref: "#/definitions/CovidCountry"
        400:
          description: "Requête invalide"
  /covid_country/{id}:
    get:
      tags: [CovidCountry]
      summary: "Récupère les données COVID pour un pays spécifique"
      description: "Retourne les détails d'une entrée 'covid_country' par son ID, incluant les noms du pays et du continent."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
      responses:
        200:
          description: "Donnée COVID d'un pays spécifique"
          schema:
            $ref: "#/definitions/CovidCountry"
        404:
          description: "Donnée non trouvée"
    put:
      tags: [CovidCountry]
      summary: "Met à jour les données COVID d'un pays spécifique"
      description: "Met à jour les détails d'une entrée 'covid_country' par son ID."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Données à mettre à jour"
          required: true
          schema:
            $ref: "#/definitions/CovidCountryInput"
      responses:
        200:
          description: "Donnée mise à jour avec succès"
        400:
          description: "Requête invalide"
        404:
          description: "Donnée non trouvée"
    delete:
      tags: [CovidCountry]
      summary: "Supprime les données COVID d'un pays spécifique"
      description: "Supprime une entrée 'covid_country' par son ID."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
      responses:
        200:
          description: "Donnée supprimée avec succès"
        404:
          description: "Donnée non trouvée"
  /monkeypox_data:
    get:
      tags: [MonkeypoxData]
      summary: "Récupère toutes les données de monkeypox"
      description: "Retourne toutes les entrées de la table 'monkeypox_data', incluant les noms du pays et du continent."
      responses:
        200:
          description: "Données monkeypox"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MonkeypoxData"
    post:
      tags: [MonkeypoxData]
      summary: "Ajoute de nouvelles données de monkeypox"
      description: "Ajoute une nouvelle entrée dans la table 'monkeypox_data'."
      parameters:
        - in: "body"
          name: "body"
          description: "Données Monkeypox à ajouter"
          required: true
          schema:
            $ref: "#/definitions/MonkeypoxDataInput"
      responses:
        201:
          description: "Donnée ajoutée avec succès"
          schema:
            $ref: "#/definitions/MonkeypoxData"
        400:
          description: "Requête invalide"
  /monkeypox_data/{id}:
    get:
      tags: [MonkeypoxData]
      summary: "Récupère les données de monkeypox pour un enregistrement spécifique"
      description: "Retourne les détails d'une entrée 'monkeypox_data' par son ID, incluant les noms du pays et du continent."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
      responses:
        200:
          description: "Donnée monkeypox spécifique"
          schema:
            $ref: "#/definitions/MonkeypoxData"
        404:
          description: "Donnée non trouvée"
    put:
      tags: [MonkeypoxData]
      summary: "Met à jour les données de monkeypox pour un enregistrement spécifique"
      description: "Met à jour les détails d'une entrée 'monkeypox_data' par son ID."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Données à mettre à jour"
          required: true
          schema:
            $ref: "#/definitions/MonkeypoxDataInput"
      responses:
        200:
          description: "Donnée mise à jour avec succès"
        400:
          description: "Requête invalide"
        404:
          description: "Donnée non trouvée"
    delete:
      tags: [MonkeypoxData]
      summary: "Supprime les données de monkeypox pour un enregistrement spécifique"
      description: "Supprime une entrée 'monkeypox_data' par son ID."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
      responses:
        200:
          description: "Donnée supprimée avec succès"
        404:
          description: "Donnée non trouvée"
  /coronavirus_daily:
    get:
      tags: [CoronavirusDaily]
      summary: "Récupère toutes les données quotidiennes de coronavirus"
      description: "Retourne toutes les entrées de la table 'coronavirus_daily', incluant les noms du pays et du continent."
      responses:
        200:
          description: "Données quotidiennes de coronavirus"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CoronavirusDaily"
    post:
      tags: [CoronavirusDaily]
      summary: "Ajoute de nouvelles données quotidiennes de coronavirus"
      description: "Ajoute une nouvelle entrée dans la table 'coronavirus_daily'."
      parameters:
        - in: "body"
          name: "body"
          description: "Données Coronavirus quotidiennes à ajouter"
          required: true
          schema:
            $ref: "#/definitions/CoronavirusDailyInput"
      responses:
        201:
          description: "Donnée ajoutée avec succès"
          schema:
            $ref: "#/definitions/CoronavirusDaily"
        400:
          description: "Requête invalide"
  /coronavirus_daily/{id}:
    get:
      tags: [CoronavirusDaily]
      summary: "Récupère les données quotidiennes de coronavirus pour un enregistrement spécifique"
      description: "Retourne les détails d'une entrée 'coronavirus_daily' par son ID, incluant les noms du pays et du continent."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
      responses:
        200:
          description: "Donnée quotidienne de coronavirus spécifique"
          schema:
            $ref: "#/definitions/CoronavirusDaily"
        404:
          description: "Donnée non trouvée"
    put:
      tags: [CoronavirusDaily]
      summary: "Met à jour les données quotidiennes de coronavirus pour un enregistrement spécifique"
      description: "Met à jour les détails d'une entrée 'coronavirus_daily' par son ID."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Données à mettre à jour"
          required: true
          schema:
            $ref: "#/definitions/CoronavirusDailyInput"
      responses:
        200:
          description: "Donnée mise à jour avec succès"
        400:
          description: "Requête invalide"
        404:
          description: "Donnée non trouvée"
    delete:
      tags: [CoronavirusDaily]
      summary: "Supprime les données quotidiennes de coronavirus pour un enregistrement spécifique"
      description: "Supprime une entrée 'coronavirus_daily' par son ID."
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
      responses:
        200:
          description: "Donnée supprimée avec succès"
        404:
          description: "Donnée non trouvée"
definitions:
  Continent:
    type: "object"
    properties:
      idContinent:
        type: "integer"
        description: "ID du continent (clé primaire)"
      continent:
        type: "string"
        description: "Nom du continent"
    required:
      - idContinent
      - continent
  Pays:
    type: "object"
    properties:
      id_pays:
        type: "integer"
        description: "ID du pays (clé primaire)"
      pays:
        type: "string"
        description: "Nom du pays"
      idContinent:
        type: "integer"
        description: "ID du continent associé"
      continent:
        type: "string"
        description: "Nom du continent (jointure, lecture seule)"
    required:
      - id_pays
      - pays
      - idContinent
  PaysInput: # Schéma pour les requêtes POST/PUT sans l'ID généré automatiquement
    type: "object"
    properties:
      id_pays:
        type: "integer"
        description: "ID du pays (clé primaire)"
      pays:
        type: "string"
        description: "Nom du pays"
      idContinent:
        type: "integer"
        description: "ID du continent associé"
    required:
      - id_pays
      - pays
      - idContinent
  CovidCountry:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "ID de l'enregistrement (clé primaire auto-générée)"
      nbCas:
        type: "integer"
        description: "Nombre de cas confirmés"
      nbMort:
        type: "integer"
        description: "Nombre de décès"
      nbSoigne:
        type: "integer"
        description: "Nombre de cas rétablis"
      nbActif:
        type: "integer"
        description: "Nombre de cas actifs"
      nbNewCas:
        type: "integer"
        description: "Nouveaux cas"
      nbNewMort:
        type: "integer"
        description: "Nouveaux décès"
      nbNewSoigne:
        type: "integer"
        description: "Nouveaux cas rétablis"
      mort100cas:
        type: "number"
        format: "float"
        description: "Taux de mortalité pour 100 cas"
      soigne100cas:
        type: "number"
        format: "float"
        description: "Taux de rétablissement pour 100 cas"
      mort100soigne:
        type: "number"
        format: "float"
        description: "Taux de mortalité pour 100 cas rétablis"
      derniereSemaineConfirme:
        type: "integer"
        description: "Cas confirmés la semaine dernière"
      changementSemaine:
        type: "integer"
        description: "Changement sur une semaine"
      pourcentageAugmentationSemaine:
        type: "number"
        format: "float"
        description: "Pourcentage d'augmentation sur une semaine"
      continent:
        type: "string"
        description: "Nom du continent (jointure, lecture seule)"
      id_pays:
        type: "integer"
        description: "ID du pays associé"
      idContinent:
        type: "integer"
        description: "ID du continent associé"
      pays:
        type: "string"
        description: "Nom du pays (jointure, lecture seule)"
    required:
      - nbCas
      - nbMort
      - nbSoigne
      - id_pays
      - idContinent
  CovidCountryInput: # Schéma pour les requêtes POST/PUT
    type: "object"
    properties:
      nbCas:
        type: "integer"
      nbMort:
        type: "integer"
      nbSoigne:
        type: "integer"
      nbActif:
        type: "integer"
      nbNewCas:
        type: "integer"
      nbNewMort:
        type: "integer"
      nbNewSoigne:
        type: "integer"
      mort100cas:
        type: "number"
        format: "float"
      soigne100cas:
        type: "number"
        format: "float"
      mort100soigne:
        type: "number"
        format: "float"
      derniereSemaineConfirme:
        type: "integer"
      changementSemaine:
        type: "integer"
      pourcentageAugmentationSemaine:
        type: "number"
        format: "float"
      continent:
        type: "string"
      id_pays:
        type: "integer"
      idContinent:
        type: "integer"
    required:
      - nbCas
      - nbMort
      - nbSoigne
      - id_pays
      - idContinent
...
  CoronavirusDailyInput: # Schéma pour les requêtes POST/PUT
    type: "object"
    properties:
      date:
        type: "string"
        format: "date"
      cumulCasTotaux:
        type: "integer"
      nouveauCasJournalier:
        type: "integer"
      casActif:
        type: "integer"
      cumulMortTotaux:
        type: "integer"
      nouvelleMortJournaliere:
        type: "integer"
      id_pays:
        type: "integer"
      idContinent:
        type: "integer"
    required:
      - date
      - cumulCasTotaux
      - nouveauCasJournalier
      - id_pays
      - idContinent